season2/requirements.txt
tensorflow-gpu==1.12.0
tqdm